ACCUMULO-4229 SyncingTabletLocator ensures caching clients align with tserver
The previously failing test case now passes.

A boolean indicator variable in TabletLocator
is set to false when the static TabletLocator cache is cleared.
Client writers and readers should check the indicator
before using a locally cached Locator.

A SyncingTabletLocator automatically fetches the current TabletLocator
when the cached TabletLocator becomes outdated (when isValid flips false).
Clients now stay in sync with the static TabletLocator cache
through the use of SyncingTabletLocator, which
TimeoutTabletLocator now extends.

Variables assumed invariant for SyncingTabletLocator are now final.
This does not affect semantics.

The new design allows for TabletServerBatchWriter to use its local Locator cache
more aggressively than it used to.

The invalidateCache() methods of SyncingTabletLocator defensively sync,
guarding against goofy code that might use more than one TabletLocator.

It may be possible to replace the static locators collection with a Guava class.
This is not done here for simplicity.

Closes apache/accumulo#96